requirements/action.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/action.in -o requirements/action.txt
across-tools @ git+ssh://git@github.com/ACROSS-Team/across-tools.git@cd87d0aadc7ec37bee5894efcd80ce896f08a5e8
    # via -r requirements/base.in
aiosmtplib==4.0.0
    # via -r requirements/base.in
alembic==1.15.1
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
argon2-cffi==25.1.0
    # via -r requirements/base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgi-correlation-id==4.3.4
    # via -r requirements/base.in
asgi-ratelimit==0.10.0
    # via -r requirements/base.in
astropy==7.0.1
    # via
    #   across-tools
    #   astroquery
    #   healpy
    #   pyvo
astropy-iers-data==0.2025.2.10.0.33.26
    # via astropy
astroquery==0.4.9.post1
    # via across-tools
asyncpg==0.30.0
    # via -r requirements/base.in
attrs==24.2.0
    # via rush
beautifulsoup4==4.13.3
    # via astroquery
boto3==1.38.45
    # via -r requirements/base.in
botocore==1.38.45
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.37.29
    # via types-boto3
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.7
    # via uvicorn
coverage==7.6.9
    # via pytest-cov
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via pydantic
fastapi==0.115.11
    # via
    #   -r requirements/base.in
    #   pytest-fastapi-deps
filelock==3.16.1
    # via virtualenv
geoalchemy2==0.17.1
    # via -r requirements/base.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
healpy==1.18.1
    # via across-tools
html5lib==1.1
    # via astroquery
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements/base.in
    #   across-tools
    #   spacetrack
identify==2.6.3
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
keyring==25.6.0
    # via astroquery
logbook==1.8.0
    # via spacetrack
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via mako
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
mypy==1.15.0
    # via -r requirements/types.in
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.0
    # via
    #   across-tools
    #   astropy
    #   astroquery
    #   healpy
    #   pyerfa
    #   shapely
    #   spiceypy
    #   types-shapely
packaging==24.2
    # via
    #   asgi-correlation-id
    #   astropy
    #   geoalchemy2
    #   pytest
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.1.0
    # via -r requirements/lint.in
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements/base.in
    #   across-tools
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.8.1
    # via
    #   -r requirements/base.in
    #   across-tools
pyerfa==2.0.1.5
    # via astropy
pyjwt==2.10.1
    # via -r requirements/base.in
pytest==8.3.5
    # via
    #   -r requirements/test.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-fastapi-deps
pytest-asyncio==0.25.3
    # via -r requirements/test.in
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-fastapi-deps==0.2.3
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.20
    # via -r requirements/base.in
pyvo==1.6.1
    # via astroquery
pyyaml==6.0.2
    # via
    #   astropy
    #   pre-commit
represent==2.1
    # via spacetrack
requests==2.32.3
    # via
    #   astroquery
    #   pyvo
ruff==0.9.6
    # via -r requirements/lint.in
rush==2021.4.0
    # via spacetrack
s3transfer==0.13.0
    # via boto3
sgp4==2.24
    # via across-tools
shapely==2.0.7
    # via
    #   -r requirements/base.in
    #   across-tools
six==1.17.0
    # via
    #   html5lib
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   spacetrack
soupsieve==2.6
    # via beautifulsoup4
spacetrack==1.3.1
    # via across-tools
spiceypy==6.0.1
    # via across-tools
sqlalchemy==2.0.37
    # via
    #   -r requirements/base.in
    #   alembic
    #   geoalchemy2
starlette==0.45.3
    # via
    #   asgi-correlation-id
    #   fastapi
structlog==25.1.0
    # via -r requirements/base.in
types-awscrt==0.25.7
    # via botocore-stubs
types-boto3==1.37.30
    # via -r requirements/types.in
types-boto3-rds==1.37.21
    # via types-boto3
types-boto3-ssm==1.37.19
    # via types-boto3
types-boto3-sts==1.37.0
    # via types-boto3
types-pyasn1==0.6.0.20240913
    # via types-python-jose
types-python-jose==3.4.0.20250224
    # via -r requirements/types.in
types-requests==2.32.0.20250306
    # via -r requirements/types.in
types-s3transfer==0.13.0
    # via types-boto3
types-shapely==2.0.0.20250202
    # via -r requirements/types.in
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.2.3
    # via
    #   botocore
    #   requests
    #   types-requests
uvicorn==0.34.0
    # via -r requirements/base.in
virtualenv==20.28.0
    # via pre-commit
webencodings==0.5.1
    # via html5lib
